backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymupdf
huggingface_hub
transformers
torch
requests
python-dotenv